// src/views/werewolf-host/types.ts
export type RoleId =
  | 'wolf'
  | 'wolf-cub'
  | 'cursed-wolf'
  | 'villager'
  | 'seer'
  | 'guard'
  | 'witch'
  | 'hunter'
  | 'disruptor'
  | 'traitor'
  | 'cupid'

export type Faction = 'wolf' | 'villager'

export type GamePhase =
  | 'setup'
  | 'role-reveal'
  | 'night'
  | 'day-result'
  | 'day-discussion'
  | 'day-nominate'
  | 'day-explain'
  | 'day-hang'
  | 'game-over'

export type NightStep =
  | 'disruptor'
  | 'cupid'
  | 'wolves'
  | 'traitor'
  | 'seer'
  | 'guard'
  | 'witch'
  | 'hunter'

export type NightUiState = 'calling' | 'acting' | 'fake-wait' | 'sleeping'

export type DeathReason = 'wolf' | 'witch-poison' | 'hunter-shot' | 'hanged' | 'lover-death'
// ...
export interface DeathRecord {
  playerId: string
  reason: DeathReason
}